[test] Move timeout test util to pkg/util/testutil
